requirements.txt
opencv-python-headless==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.0.1

